# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/teams": {
      "get": {
        "description": "Admins see every team. Everyone else sees only the teams they belong to.",
        "operationId": "getTeams",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Teams."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List teams",
        "tags": [
          "Teams"
        ]
      },
      "post": {
        "description": "Admins create parent teams for the organisation.",
        "operationId": "postTeams",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Team created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{id}": {
      "get": {
        "description": "Returns a team with its sub-teams, phases, and members.",
        "operationId": "getTeamsById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Team id",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Team detail."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a team",
        "tags": [
          "Teams"
        ]
      },
      "patch": {
        "description": "Admins and the team's lead can rename the team.",
        "operationId": "patchTeams",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Team id",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Team updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{id}/sub-teams": {
      "post": {
        "description": "Admins and team leads can add a sub-team under a parent team.",
        "operationId": "postTeamsSubTeams",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Team id",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sub-team created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a sub-team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{id}/phases": {
      "put": {
        "description": "Sets the ordered pipeline for new projects. Phases that share a step run in parallel. Existing projects keep the pipeline they started with.",
        "operationId": "putTeamsPhases",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Team id",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplacePhasesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Phases saved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace a team's phases",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{id}/invites": {
      "get": {
        "description": "Admins and team leads can see outstanding and accepted invites.",
        "operationId": "getTeamsInvites",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Team id",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Invites."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List invites for a team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/invites": {
      "post": {
        "description": "Admins can invite anyone to any team and sub-team, including as a lead. Team leads can invite members to their own team and assign a sub-team. Plugin emails the invite when Resend is configured, and always returns a copyable invite URL.",
        "operationId": "postInvites",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invite created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite a user to a team",
        "tags": [
          "Teams",
          "Invites"
        ]
      }
    },
...
  },
  "info": {
    "title": "Plugin Project Manager API",
    "description": "API for Plugin teams to create projects, move work through configurable phases, and hand off documents.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Production"
          },
          "slug": {
            "type": "string",
            "example": "production"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "CreateSubTeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Design"
          },
          "slug": {
            "type": "string",
            "example": "design"
          }
        },
        "required": [
          "name"
        ]
      },
      "PhaseItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Design"
          },
          "slug": {
            "type": "string",
            "example": "design",
            "description": "Generated from the phase name if omitted."
          },
          "step": {
            "type": "number",
            "example": 1,
            "description": "Phases that share a step run in parallel. The next step starts when every phase on this step is done."
          },
          "subTeamId": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "step",
          "subTeamId"
        ]
      },
      "ReplacePhasesDto": {
        "type": "object",
        "properties": {
          "phases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhaseItemDto"
            }
          }
        },
        "required": [
          "phases"
        ]
      },
      "CreateInviteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ama@plugin.local"
          },
          "teamId": {
            "type": "string"
          },
          "subTeamId": {
            "type": "string"
          },
          "teamRole": {
            "type": "string",
            "enum": [
              "lead",
              "member"
            ],
            "default": "member"
          }
        },
        "required": [
          "email",
          "teamId"
        ]
      },
...
    }
  }
}